oos: purity-agnostic and purity-only filters, and whole-packet processors

+++ apps/outofspace/desktop/src/main/kotlin/OutofspaceMain.kt
@@ -271,11 +271,11 @@ fun main() {
             GLFW_KEY_F10 -> hud.saveStatus = loadWorld(controller, renderer)
             GLFW_KEY_C -> {
                 val (ix, iy) = renderer.screenToTile(lastX, lastY)
-                hud.clipboardStatus = controller.copySettings(TileIndex(ix + iy * controller.state.grid.width))
+                hud.clipboardStatus = controller.copySettings(TileIndex(ix.toInt() + iy.toInt() * controller.state.grid.width))
             }
             GLFW_KEY_V -> {
                 val (ix, iy) = renderer.screenToTile(lastX, lastY)
-                hud.clipboardStatus = controller.pasteSettings(TileIndex(ix + iy * controller.state.grid.width))
+                hud.clipboardStatus = controller.pasteSettings(TileIndex(ix.toInt() + iy.toInt() * controller.state.grid.width))
             }
             GLFW_KEY_LEFT_BRACKET -> controller.speed = max(0.25f, controller.speed / 2f)
             GLFW_KEY_RIGHT_BRACKET -> controller.speed = (controller.speed * 2f).coerceAtMost(16f)
